Initial sprites for deer, armadillo, boars and badgers.

+++ src/raws/creatures.hpp
@@ -42,7 +42,7 @@ struct raw_creature_t {
     int yield_bone=0;
     int yield_skull=0;
     creature_ai_t ai;
-    uint8_t glyph;
+    uint16_t glyph;
     rltk::color_t fg;
     int hp_n, hp_dice, hp_mod;
     int group_size_n_dice, group_size_dice, group_size_mod;
